+++ packages/e2e-utils/src/index.ts
@@ -2,6 +2,7 @@ export { BackendWebsocketServerMock } from './mocks/backendServer';
 export { TrezorBridgeMock } from './mocks/bridge';
 export { DropboxMock } from './mocks/dropbox';
 export { GoogleMock } from './mocks/google';
+export { GitHubReporterBase, InitializationState } from './githubReporter/gitHubReporterBase';
 export { GitHubProject } from './githubReporter/gitHubProject';
 export { IssueRequests } from './githubReporter/issueRequests';
 export * from './githubReporter/types';
