--- suite-native/app/e2e/support/testAnnotations.ts
@@ -1,19 +1,0 @@
-export enum TestCategory {
-  Settings = 'Settings',
-  Transactions = 'Transactions',
-  Accounts = 'Accounts',
-  Dashboard = 'Dashboard',
-  Guide = 'Guide',
-  OSFeatureLink = 'OS feature link',
-  SplashScreen = 'Splash screen',
-  ImportProcess = 'Import process',
-  Device = 'Device',
-}
-
-export enum TestPriority {
-    Critical = 'Critical',
-    High = 'High',
-    Medium = 'Medium',
-    Low = 'Low',
-    AsNecessary = 'As necessary',
-}
